requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
SQLAlchemy==2.0.30
pydantic==2.7.2
python-dotenv==1.0.1
langchain==0.2.7
langchain-community==0.2.7
langchain-openai==0.1.7
faiss-cpu==1.8.0
sentence-transformers==3.0.1
requests==2.32.3
# MCP server (Model Context Protocol)
mcp==1.1.2
